Add inconsistent municipalities
We add a new script to the data processing pipeline that produces data on which municipalities are inconsistent across years. Just in case the set of missing years is not 'contiguous' (this was the case with Grande-Anse prior to bugfixes in PRs #1 and #2, after all), we save not just the first and last years but rather all the missing years for each relevant municipality as a list.