Update Florenceville-Bristol calculations (#5)
The fiscal capacity computations when combining Florenceville and Bristol prior to 2009 were flawed (the true formula is extremely obscure); since we have no need for them in our data analysis, we simply dropped them entirely. Moreover, we also fixed the flawed average tax rate computations for the amalgamation, diving warrant by total tax base (for rate).

This commit thus updates the script and reruns it to obtain the corrected spreadsheets.